Postgres testcontainers running with init file and skunk

+++ http-server/src/main/scala/httpserver/domain/GpsCoords.scala
@@ -1,6 +1,6 @@
 package httpserver.domain
 
 case class GpsCoords(
-    lat: Double,
-    long: Double,
+    lat: BigDecimal,
+    long: BigDecimal,
 )
